+++ src/context/ContextType.ts
@@ -9,10 +9,11 @@ export enum ContextTypes {
   Email = 'fdc3.email',
   Instrument = 'fdc3.instrument',
   InstrumentList = 'fdc3.instrumentList',
+  Nothing = 'fdc3.nothing',
   Organization = 'fdc3.organization',
   Portfolio = 'fdc3.portfolio',
   Position = 'fdc3.position',
-  Nothing = 'fdc3.nothing',
+  ChatSearchCriteria = 'fdc3.chat.searchCriteria',
   TimeRange = 'fdc3.timerange',
   Valuation = 'fdc3.valuation',
 }