+++ src/context/ContextTypes.ts
@@ -187,6 +187,11 @@ export interface Position {
   name?: string;
 }
 
+export interface ChatSearchCriteria {
+  criteria: (Instrument | Organization | Contact | string)[];
+  type: string;
+}
+
 export interface Valuation {
   CURRENCY_ISOCODE: string;
   expiryTime?: Date;