+++ src/intents/Intents.ts
@@ -10,6 +10,7 @@ export enum Intents {
   ViewHoldings = 'ViewHoldings',
   ViewInstrument = 'ViewInstrument',
   ViewInteractions = 'ViewInteractions',
+  ViewMessages = 'ViewMessages',
   ViewNews = 'ViewNews',
   ViewOrders = 'ViewOrders',
   ViewProfile = 'ViewProfile',